requirements.txt
strands-agents
boto3
PyPDF2
pydantic
uuid
